editoast: create bouding_box.rs in primitives folder
